+++ lib/src/inbound/src/protocol.dart
@@ -104,6 +104,9 @@ class InboundProtocol with _$InboundProtocol {
 
     /// [hasModbus] is the [bool] value that indicates if the protocol has support for Modbus.
     bool? hasModbus,
+
+    /// [modbusPorts] is the list of Modbus ports that the protocol has.
+    @Default([]) List<String> modbusPorts,
   }) = _InboundProtocol;
 
   factory InboundProtocol.fromJson(Map<String, dynamic> json) => _$InboundProtocolFromJson(json);