+++ lib/src/modbus/modbus.dart
@@ -1,0 +1,11 @@
+library modbus;
+
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+import '../../layrz_models.dart';
+
+part 'modbus.freezed.dart';
+part 'modbus.g.dart';
+
+part 'src/config.dart';
+part 'src/parameter.dart';